+++ CHANGELOG.md
@@ -1,6 +1,28 @@
 ##Changelog
 
-###v1.2.0 (2013-06-xx)
+###v1.2.5 (2013-10-30)
+
+* [fix] Fix wrong method name
+
+###v1.2.4 (2013-10-30)
+
+* [new] Don't use `sudo` to start workers if the current user is already the target user
+
+###v1.2.3 (2013-10-26)
+
+* [new] Fix #21: Add new tmpdir to fresque.ini to specify where to save the PID files [@pwhelan]
+
+###v1.2.2 (2013-10-23)
+
+* [fix] Fix #20: the bin script inside composer bin folder can't find library files
+
+###v1.2.1 (2013-06-08)
+
+* [new] `load` command now supports starting workers polling multiple queues
+* [fix] Fix bug where multiple workers will not start correctly
+* [fix] Fix load command: each queues should run under its own settings
+
+###v1.2.0 (2013-06-20)
 
 * [new] Add `pause` and `resume` commands
 * [new] Add `--debug` option